feat: impl clubs student frame
