+++ packages/web/src/app/clubs/page.tsx
@@ -1,21 +1,15 @@
 "use client";
 
-import React, { useEffect, useMemo, useState } from "react";
-
-import { RegistrationDeadlineEnum } from "@sparcs-clubs/interface/common/enum/registration.enum";
-import { hangulIncludes } from "es-hangul";
+import React from "react";
 
 import styled from "styled-components";
 
-import AsyncBoundary from "@sparcs-clubs/web/common/components/AsyncBoundary";
 import FlexWrapper from "@sparcs-clubs/web/common/components/FlexWrapper";
-import Info from "@sparcs-clubs/web/common/components/Info";
 import PageHead from "@sparcs-clubs/web/common/components/PageHead";
-import SearchInput from "@sparcs-clubs/web/common/components/SearchInput";
-import ClubsSectionFrame from "@sparcs-clubs/web/features/clubs/frames/ClubsSectionFrame";
-import { useGetClubsList } from "@sparcs-clubs/web/features/clubs/services/useGetClubsList";
-import { useGetRegistrationTerm } from "@sparcs-clubs/web/features/clubs/services/useGetRegistrationTerm";
-import { formatDateTime } from "@sparcs-clubs/web/utils/Date/formatDate";
+import { useAuth } from "@sparcs-clubs/web/common/providers/AuthContext";
+import ClubsListFrame from "@sparcs-clubs/web/features/clubs/frames/ClubsListFrame";
+import ClubsStudentFrame from "@sparcs-clubs/web/features/clubs/frames/ClubsStudentFrame";
+import isStudent from "@sparcs-clubs/web/utils/isStudent";
 
 const ResponsiveWrapper = styled(FlexWrapper)`
   gap: 60px;
@@ -27,63 +21,7 @@ const ResponsiveWrapper = styled(FlexWrapper)`
 `;
 
 const Clubs: React.FC = () => {
-  const { data, isLoading, isError } = useGetClubsList();
-  const {
-    data: termData,
-    isLoading: isLoadingTerm,
-    isError: isErrorTerm,
-  } = useGetRegistrationTerm();
-  const [isRegistrationPeriod, setIsRegistrationPeriod] = useState<boolean>();
-  const [memberRegistrationPeriodEnd, setMemberRegistrationPeriodEnd] =
-    useState<Date>(new Date());
-  const [searchText, setSearchText] = useState<string>("");
-
-  useEffect(() => {
-    if (termData) {
-      const now = new Date();
-      const currentEvents = termData.events.filter(
-        event => now >= event.startTerm && now <= event.endTerm,
-      );
-      if (currentEvents.length === 0) {
-        setIsRegistrationPeriod(false);
-        return;
-      }
-      const registrationEvent = currentEvents.filter(
-        event =>
-          event.registrationEventEnumId ===
-          RegistrationDeadlineEnum.StudentRegistrationApplication,
-      );
-      if (registrationEvent.length > 0) {
-        setIsRegistrationPeriod(true);
-        setMemberRegistrationPeriodEnd(registrationEvent[0].endTerm);
-      } else {
-        setIsRegistrationPeriod(false);
-      }
-    }
-  }, [termData]);
-
-  const filteredDivisions = useMemo(
-    () =>
-      (data?.divisions ?? [])
-        .map(division => {
-          const filteredClubs = division.clubs
-            .filter(
-              item =>
-                item.name_kr.toLowerCase().includes(searchText.toLowerCase()) ||
-                item.name_en.toLowerCase().includes(searchText.toLowerCase()) ||
-                hangulIncludes(item.name_kr, searchText),
-            )
-            .sort((a, b) => {
-              if (a.isPermanent && !b.isPermanent) return -1;
-              if (!a.isPermanent && b.isPermanent) return 1;
-              return a.type - b.type || a.name_kr.localeCompare(b.name_kr);
-            });
-
-          return { ...division, clubs: filteredClubs };
-        })
-        .filter(division => division.clubs.length > 0),
-    [data, searchText],
-  );
+  const { isLoggedIn, profile } = useAuth();
 
   return (
     <ResponsiveWrapper direction="column" gap={60}>
@@ -91,32 +29,11 @@ const Clubs: React.FC = () => {
         items={[{ name: "동아리 목록", path: "/clubs" }]}
         title="동아리 목록"
       />
-      <AsyncBoundary isLoading={isLoadingTerm} isError={isErrorTerm}>
-        {isRegistrationPeriod && (
-          // TODO: 학기 동적처리
-          <Info
-            text={`현재는 2024년 가을학기 동아리 신청 기간입니다 (신청 마감 : ${formatDateTime(memberRegistrationPeriodEnd)})`}
-          />
-        )}
-      </AsyncBoundary>
-
-      <SearchInput
-        searchText={searchText}
-        handleChange={setSearchText}
-        placeholder="동아리 이름으로 검색하세요"
-      />
-      <AsyncBoundary isLoading={isLoading} isError={isError}>
-        <FlexWrapper direction="column" gap={40}>
-          {filteredDivisions.map(division => (
-            <ClubsSectionFrame
-              title={division.name}
-              clubList={division.clubs}
-              key={division.name}
-              isRegistrationPeriod={isRegistrationPeriod}
-            />
-          ))}
-        </FlexWrapper>
-      </AsyncBoundary>
+      {isLoggedIn && isStudent(profile) ? (
+        <ClubsStudentFrame />
+      ) : (
+        <ClubsListFrame isRegistrationPeriod={false} />
+      )}
     </ResponsiveWrapper>
   );
 };
