+++ packages/web/src/features/clubDetails/frames/ClubDetailInfoFrame.tsx
@@ -8,14 +8,9 @@ import FlexWrapper from "@sparcs-clubs/web/common/components/FlexWrapper";
 import SectionTitle from "@sparcs-clubs/web/common/components/SectionTitle";
 import ClubDetailCard from "@sparcs-clubs/web/features/clubDetails/components/ClubDetailCard";
 import ClubInfoCard from "@sparcs-clubs/web/features/clubDetails/components/ClubInfoCard";
-import PersonInfoCard from "@sparcs-clubs/web/features/clubDetails/components/PersonInfoCard";
-
-import type { ApiClb002ResponseOK } from "@sparcs-clubs/interface/api/club/endpoint/apiClb002";
-
-export interface ClubDetailFrameProps {
-  club: ApiClb002ResponseOK;
-  isRegistrationPeriod: boolean;
-}
+import PersonInfoCard, {
+  ClubDetailFrameProps,
+} from "@sparcs-clubs/web/features/clubDetails/components/PersonInfoCard";
 
 const CardWrapper = styled.div`
   width: 100%;
