+++ packages/web/src/features/clubs/frames/ClubsStudentFrame.tsx
@@ -1,0 +1,63 @@
+"use client";
+
+import React, { useEffect, useState } from "react";
+
+import { RegistrationDeadlineEnum } from "@sparcs-clubs/interface/common/enum/registration.enum";
+
+import AsyncBoundary from "@sparcs-clubs/web/common/components/AsyncBoundary";
+import Info from "@sparcs-clubs/web/common/components/Info";
+import ClubsListFrame from "@sparcs-clubs/web/features/clubs/frames/ClubsListFrame";
+import { useGetRegistrationTerm } from "@sparcs-clubs/web/features/clubs/services/useGetRegistrationTerm";
+import { formatDateTime } from "@sparcs-clubs/web/utils/Date/formatDate";
+
+const ClubsStudentFrame: React.FC = () => {
+  const {
+    data: termData,
+    isLoading: isLoadingTerm,
+    isError: isErrorTerm,
+  } = useGetRegistrationTerm();
+  const [isRegistrationPeriod, setIsRegistrationPeriod] =
+    useState<boolean>(false);
+  const [memberRegistrationPeriodEnd, setMemberRegistrationPeriodEnd] =
+    useState<Date>(new Date());
+
+  useEffect(() => {
+    if (termData) {
+      const now = new Date();
+      const currentEvents = termData.events.filter(
+        event => now >= event.startTerm && now <= event.endTerm,
+      );
+      if (currentEvents.length === 0) {
+        setIsRegistrationPeriod(false);
+        return;
+      }
+      const registrationEvent = currentEvents.filter(
+        event =>
+          event.registrationEventEnumId ===
+          RegistrationDeadlineEnum.StudentRegistrationApplication,
+      );
+      if (registrationEvent.length > 0) {
+        setIsRegistrationPeriod(true);
+        setMemberRegistrationPeriodEnd(registrationEvent[0].endTerm);
+      } else {
+        setIsRegistrationPeriod(false);
+      }
+    }
+  }, [termData]);
+
+  return (
+    <>
+      <AsyncBoundary isLoading={isLoadingTerm} isError={isErrorTerm}>
+        {isRegistrationPeriod && (
+          // TODO: 학기 동적처리
+          <Info
+            text={`현재는 2024년 가을학기 동아리 신청 기간입니다 (신청 마감 : ${formatDateTime(memberRegistrationPeriodEnd)})`}
+          />
+        )}
+      </AsyncBoundary>
+      <ClubsListFrame isRegistrationPeriod={isRegistrationPeriod} />
+    </>
+  );
+};
+
+export default ClubsStudentFrame;
